requirements.txt
aiohttp==3.5.4
aiohttp-jinja2==1.1.2
aiohttp-security==0.4.0
async-timeout==3.0.1
asyncpg==0.18.3
attrs==19.1.0
bcrypt==3.1.6
cffi==1.12.3
chardet==3.0.4
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
multidict==4.5.2
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
PyYAML==5.1.1
six==1.12.0
trafaret==1.2.0
yarl==1.3.0
